+++ mobile/lib/service_locator.dart
@@ -11,7 +11,7 @@ import "package:photos/services/machine_learning/face_ml/face_recognition_servic
 import "package:photos/services/machine_learning/machine_learning_controller.dart";
 import "package:photos/services/magic_cache_service.dart";
 import "package:photos/services/storage_bonus_service.dart";
-import "package:photos/services/trash_sync_service.dart";
+import "package:photos/services/sync/trash_sync_service.dart";
 import "package:photos/services/update_service.dart";
 import "package:photos/utils/local_settings.dart";
 import "package:shared_preferences/shared_preferences.dart";
