+++ mobile/lib/services/file_magic_service.dart
@@ -16,7 +16,7 @@ import "package:photos/models/api/metadata.dart";
 import 'package:photos/models/file/file.dart';
 import "package:photos/models/metadata/common_keys.dart";
 import "package:photos/models/metadata/file_magic.dart";
-import 'package:photos/services/remote_sync_service.dart';
+import 'package:photos/services/sync/remote_sync_service.dart';
 import "package:photos/utils/file_key.dart";
 
 class FileMagicService {
