+++ mobile/lib/services/sync/remote_sync_service.dart
@@ -28,7 +28,7 @@ import 'package:photos/services/collections_service.dart';
 import 'package:photos/services/ignored_files_service.dart';
 import 'package:photos/services/local_file_update_service.dart';
 import "package:photos/services/notification_service.dart";
-import 'package:photos/services/sync_service.dart';
+import 'package:photos/services/sync/sync_service.dart';
 import 'package:photos/utils/diff_fetcher.dart';
 import 'package:photos/utils/file_uploader.dart';
 import 'package:photos/utils/file_util.dart';
@@ -308,7 +308,6 @@ class RemoteSyncService {
   Future<void> joinAndSyncCollection(
     BuildContext context,
     int collectionID,
-  
   ) async {
     await _collectionsService.joinPublicCollection(context, collectionID);
     await _collectionsService.sync();
