+++ mobile/lib/ui/components/home_header_widget.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import "package:logging/logging.dart";
 import "package:photo_manager/photo_manager.dart";
 import "package:photos/generated/l10n.dart";
-import "package:photos/services/local_sync_service.dart";
+import "package:photos/services/sync/local_sync_service.dart";
 import 'package:photos/ui/components/buttons/icon_button_widget.dart';
 import "package:photos/ui/settings/backup/backup_folder_selection_page.dart";
 import "package:photos/utils/dialog_util.dart";
