+++ mobile/lib/ui/settings/debug/debug_section_widget.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:photos/core/configuration.dart';
 import "package:photos/service_locator.dart";
 import 'package:photos/services/ignored_files_service.dart';
-import 'package:photos/services/local_sync_service.dart';
-import 'package:photos/services/sync_service.dart';
+import 'package:photos/services/sync/local_sync_service.dart';
+import 'package:photos/services/sync/sync_service.dart';
 import 'package:photos/theme/ente_theme.dart';
 import 'package:photos/ui/components/captioned_text_widget.dart';
 import 'package:photos/ui/components/expandable_menu_item_widget.dart';
