+++ mobile/lib/utils/delete_file_util.dart
@@ -18,8 +18,8 @@ import 'package:photos/models/file/file.dart';
 import "package:photos/models/files_split.dart";
 import 'package:photos/models/selected_files.dart';
 import "package:photos/service_locator.dart";
-import 'package:photos/services/remote_sync_service.dart';
-import 'package:photos/services/sync_service.dart';
+import 'package:photos/services/sync/remote_sync_service.dart';
+import 'package:photos/services/sync/sync_service.dart';
 import 'package:photos/ui/common/linear_progress_dialog.dart';
 import 'package:photos/ui/components/action_sheet_widget.dart';
 import 'package:photos/ui/components/buttons/button_widget.dart';
