+++ mobile/lib/utils/file_uploader.dart
@@ -36,9 +36,9 @@ import "package:photos/module/upload/service/multipart.dart";
 import "package:photos/service_locator.dart";
 import "package:photos/services/account/user_service.dart";
 import 'package:photos/services/collections_service.dart';
-import 'package:photos/services/local_sync_service.dart';
 import "package:photos/services/preview_video_store.dart";
-import 'package:photos/services/sync_service.dart';
+import 'package:photos/services/sync/local_sync_service.dart';
+import 'package:photos/services/sync/sync_service.dart';
 import 'package:photos/utils/data_util.dart';
 import "package:photos/utils/exif_util.dart";
 import "package:photos/utils/file_key.dart";
